Breakdown config into API and SPI - provide a yaml config module - provide a toml config module - change single JAR generation, use shade plugin instead
