+++ layrry-config-toml/pom.xml
@@ -26,19 +26,19 @@
 
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>layrry-platform</artifactId>
+  <artifactId>layrry-config-toml</artifactId>
   <packaging>jar</packaging>
-  <name>Layrry Platform</name>
+  <name>TOML config support for Layrry</name>
 
   <dependencies>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>layrry-config</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.github.jezza</groupId>
+      <artifactId>toml</artifactId>
+      <version>1.2</version>
+    </dependency>
   </dependencies>
-
-  <build>
-    <pluginManagement>
-      <plugins>
-      </plugins>
-    </pluginManagement>
-    <plugins>
-    </plugins>
-  </build>
 </project>
