+++ layrry-config-yaml/pom.xml
@@ -26,19 +26,20 @@
 
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>layrry-platform</artifactId>
+  <artifactId>layrry-config-yaml</artifactId>
   <packaging>jar</packaging>
-  <name>Layrry Platform</name>
+  <name>YAML config support for Layrry</name>
 
   <dependencies>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>layrry-config</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.yaml</groupId>
+      <artifactId>snakeyaml</artifactId>
+      <version>1.26</version>
+    </dependency>
   </dependencies>
 
-  <build>
-    <pluginManagement>
-      <plugins>
-      </plugins>
-    </pluginManagement>
-    <plugins>
-    </plugins>
-  </build>
 </project>
