+++ layrry-config-yaml/src/main/resources/META-INF/services/org.moditect.layrry.config.LayersConfigParser
@@ -1,0 +1,1 @@
+org.moditect.layrry.config.YamlLayersConfigParser
