+++ layrry-config/pom.xml
@@ -26,19 +26,8 @@
 
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>layrry-platform</artifactId>
+  <artifactId>layrry-config</artifactId>
   <packaging>jar</packaging>
-  <name>Layrry Platform</name>
+  <name>Layrry Config</name>
 
-  <dependencies>
-  </dependencies>
-
-  <build>
-    <pluginManagement>
-      <plugins>
-      </plugins>
-    </pluginManagement>
-    <plugins>
-    </plugins>
-  </build>
 </project>
