+++ layrry-config/src/main/java/org/moditect/layrry/config/Layer.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.moditect.layrry.internal.descriptor;
+package org.moditect.layrry.config;
 
 import java.util.ArrayList;
 import java.util.List;
