+++ pom.xml
@@ -34,6 +34,9 @@
 
   <modules>
     <module>layrry-platform</module>
+    <module>layrry-config</module>
+    <module>layrry-config-yaml</module>
+    <module>layrry-config-toml</module>
     <module>layrry-launcher</module>
     <module>integration-test</module>
     <module>plugin-example</module>
@@ -45,6 +48,21 @@
       <dependency>
         <groupId>${project.groupId}</groupId>
         <artifactId>layrry-platform</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>${project.groupId}</groupId>
+        <artifactId>layrry-config</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>${project.groupId}</groupId>
+        <artifactId>layrry-config-yaml</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>${project.groupId}</groupId>
+        <artifactId>layrry-config-toml</artifactId>
         <version>${project.version}</version>
       </dependency>
       <dependency>
